Split onboarding folder startup repo test
